requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
requests==2.31.0
beautifulsoup4==4.12.3
feedparser==6.0.11
anthropic==0.34.0
python-dotenv==1.0.1
APScheduler==3.10.4
PyJWT==2.8.0
google-auth==2.25.2
email-validator==2.1.0.post1
sendgrid==6.11.0
jinja2==3.1.2
premailer==3.10.0